requirements.txt
Django
djangorestframework
python-decouple
djangorestframework-simplejwt
celery
redis
python-telegram-bot
